Added project files to new repo
I've tranferred all my project 1 files to this repo because I couldn't figure out how to rename the old repo I was working with. I never changed the default "portfolio_update_v1" name given to the assets file on Treehouse and I want to give the project an accurate title.